fikser tester for brevredigering & fiks bug ved navigering mellom brevredigering & brevvelger
